fixed get route in comment-route

+++ controllers/api/comment-routes.js
@@ -2,17 +2,20 @@ const router = require('express').Router();
 const { Comment } = require('../../models');
 
 router.get('/', (req, res) => {
-
+    Comment.findAll()
+        .then(dbCommentData => res.json(dbCommentData))
+        .catch(err => {
+            console.log(err);
+            res.status(500).json(err);
+        });
 });
 
 router.post('/', (req, res) => {
-    // check the session
     if (req.session) {
         Comment.create({
             comment_text: req.body.comment_text,
-            post_id: req.body.post_id,
-            // use the id from the session
-            user_id: req.session.user_id
+            user_id: req.session.user_id,
+            post_id: req.body.post_id
         })
             .then(dbCommentData => res.json(dbCommentData))
             .catch(err => {
@@ -23,22 +26,24 @@ router.post('/', (req, res) => {
 });
 
 router.delete('/:id', (req, res) => {
-    Comment.destroy({
-        where: {
-            id: req.params.id
-        }
-    })
-        .then(dbCommentData => {
-            if (!dbCommentData) {
-                res.status(404).json({ message: 'No comment found with this id!' });
-                return;
+    if (req.session) {
+        Comment.destroy({
+            where: {
+                id: req.params.id
             }
-            res.json(dbCommentData);
         })
-        .catch(err => {
-            console.log(err);
-            res.status(500).json(err);
-        });
+            .then(dbCommentData => {
+                if (!dbCommentData) {
+                    res.status(404).json({ message: 'No comment found with this id!' });
+                    return;
+                }
+                res.json(dbCommentData);
+            })
+            .catch(err => {
+                console.log(err);
+                res.status(500).json(err);
+            });
+    }
 });
 
 module.exports = router;
